Revert "Revert "Minor Refactoring to put all relevant input in their respective classes""
This reverts commit 67f6fde360e1d4da0a2064234dfb6c4be7261ce0.
