+++ Finishers/Assets/Characters/Player/Player.prefab
@@ -1956,7 +1956,7 @@ GameObject:
   - component: {fileID: 3706741003250733307}
   - component: {fileID: 3079736054807551836}
   - component: {fileID: 3079736054807551835}
-  - component: {fileID: 1962851768}
+  - component: {fileID: 7070460760452883848}
   - component: {fileID: 7146908285708924571}
   - component: {fileID: 4620075608317267798}
   m_Layer: 12
@@ -2051,7 +2051,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 6009fa1d308ae9a44884b0f1f1b46226, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &1962851768
+--- !u!114 &7070460760452883848
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -2060,7 +2060,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 3079736054807551833}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 7059047519f571448b9bc379b53f89c3, type: 3}
+  m_Script: {fileID: 11500000, guid: cd70edd967335ec4b8da63f269c76bda, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   config: {fileID: 11400000, guid: dacdbedf25cb96847acf7fc3ba8ca9fb, type: 2}
